PoolingContainer library
This adds an androidx.customview:customview-poolingcontainer artifact that is
depended on by both Compose and RecyclerView, through which AbstractComposeView
and RecyclerView discover each other and communicate about when the Compose
view should dispose its composition.

This mechanism is independent of both Compose and RecyclerView, and
could be used for any recycling container or child with heavy resources
that should be retained across recycling.

Relnote: "Add a new PoolingContainer library that allows for listening
to dispose events of a container that manages its children outside the View
hierarchy.  This will later be added as a dependency of Compose and RecyclerView"
Test: New PoolingContainerTest
Bug: 196371929

Change-Id: I0e3f61f2a38f28a3fa9e5cbd50dbfb10c0871984