Avoid unnecessary dismiss current fragment with vertical fling gesture
In case of SwipeDismissFrameLayout fragment with non-scrollable vertically,
if user tries fling gesture upward, it can returns to the previous fragment.
This issue occurs even if user does not intend to dismiss current screen
with vertical fling gesture.

The reason why this issue occurs is that there is no logic to check
if X velocity value is greater than Y velocity value. To avoid unintended
going back, we should add logic to check X velocity value is greater that Y’s one.

Bug: 196928100
Test: Manually running sample APK.

Change-Id: Idb6d8a39b1b2d5760d1f5e9ec9678dc75d15dde3
Signed-off-by: Dae Gyu LEE <de_amour.lee@samsung.com>