Lazily resolve raw types
This CL changes KSPRawType to lazily calculate typename instead of
eagerly computing it.

I've also changed KspProcessingEnv to cache arrayTypeFactory (showed in
profiles).

To validate this, I've updated profile.sh to support test runs. It will
be merged in a followup CL as we are likely to bikeshed there :).
In the profile test, this removed a ~25% time spent while getting the
raw type out of a type element (typeElement.type.rawType).

Bug: 204923211
Test: profile
Change-Id: I792f643a80294a65bf55572cc4872b4a8af4e1b7