Stop using star projection in KSP types
When converting a declaration to type, we've always used star projection
in KSP which loses type parameter information.

This CL changes it to just use KSClassDeclaration.asType(emptylist()) instead
of KSClassDeclaration.starProjection(). It also brings KSP and KAPT closer.
(see updated tests that used to special case KSP).

Note that, this only fixes part of the reported bug. When generating Java
signatures from Kotlin code, Kotlin has some special rules for wildcards that
we handle in OverrideVarianceResolver. Doing this makes it work for Room's use
case but also provides inconsistent type names when they are obtained from parameters.

We need to tackle that separately and decide whether we want to infer variance
while resolving those types OR do it while generating TypeNames. We'll tackle it
in a followup CL.

We might also be able to get rid of KspTypeArgumentType but I'll leave it to a followup
after we figure out the variance part.

Bug: 204415667
Test: XTypeTest
Change-Id: Ie5dc717a96a97d266506750dce4bd1c426af194f