Handle type variance for all parameter types
This CL fixes a bug in XProcessing where the TypeName for parameters did
not include variance information.
https://kotlinlang.org/docs/reference/java-to-kotlin-interop.html#variant-generics

For Room, we only needed this information while generating overrides
hence it was handled in MethodSpecHelper / OverrideVarianceResolver.
This created an inconsistent state when the TypeName is directly
obtained from the parameter.

We had 2 possible options:
a) Keep the location where we obtained the type along with the XType such
that we can change the TypeName generation
b) Update the XType returned from these parameters to include variance.

In most similar cases, we took route B to ease the migration to
XProcessing (e.g. suspend functions generate a synthetic parameter).
Given that the migration is complete and the future of Room is Kotlin, I
would rather keep the correct KSType information as Kotlin sees it (even
though it is inconsistent w/ KAPT).

I've tried multiple prototypes on how this information can be embedded
into KspType but unfortunately none are clean. The problem is that, we
have many places in the code where we obtain a KspType and we don't
always have the original location information. It would be possible to
pass it around but it creates a lot of noise.

Instead of doing that, I decided to add a new JvmTypeResolver property
to KspType which gives ability to hook into the TypeName generation. As
such, now each KspTypeElement class implements a resolveTypeName method
instead of overriding the property (which became final in KspType).

Unfortunately, I couldn't find a decent documentation on how this works
in all cases. So most of the logic in KspJvmTypeResolver /
KSTypeVarianceResolver is reverse engineered from the
KspTypeNamesGoldenTest. There is still cases where we miss for return
types but it also does not impact code generation.
Eventaully, we should be able to replace KSJvmTypeResolver with a KSP
API once it is available. https://github.com/google/ksp/issues/717

Bug: 204415667
Test: KSTypeExtTest, KspTypeNamesGoldenTest
Change-Id: I4294dcd89e3c4d4f58b3b71f7923d3223876f9c1