Add property weights to SearchSpec
This change allows clients to promote and demote properties for a schema
type when scoring documents. For instance, if a client wants to
prefer matches in the "subject" property over the "body" property of
an "Email" schema type, they can assign "subject" a higher weight
than "body".

Test: ./gradlew appsearch:appsearch:connectedCheck
 appsearch:appsearch-platform-storage:connectedCheck
 appsearch:appsearch-local-storage:connectedCheck
Bug: 203700301
Relnote: "This change adds an api to allow clients to specify property
weights to control how matches to query terms in different properties
affects the scored produced by RANKING_STRATEGY_RELEVANCE_SCORING."

Change-Id: I069b9bf88a9bd6b3b4b6fcee4e0a8ec526b71c98