Add support for synchronous complication providers
Also fixes a bug where assertions intended for preview
data where applied to non-preview data.

Bug: 202019175
Bug: 204927992
Test: Manual testing
Relnote: We've added support for synchronous complication providers where the complication is updated at a higher frequency than normal, up to once per second when the watch face is visible and non-ambient. To use this the provider must include a new androidx.wear.watchface.complications.data.source.SYNCHRONOUS_UPDATE_PERIOD_SECONDS meta data tag in it's manifest and override onSynchronousComplicationRequest. Depending on the nature of the data source, it may also need to override onStartSynchronousComplicationRequests and onStopInteractiveComplicationRequests to get notifications of when the complication enters and exits interactive mode.
Change-Id: I8fe9dcbd9b4a3a293322b69aa91c4c3a25325c63