Move drawing into an entity of LayoutNodeWrapper
Bug: 181981276

This is the first CL rearranging subclasses of LayoutNodeWrapper
to children hanging off of LayoutNodeWrapper instead of having
them as elements in the chain. This should improve performance
of calls that are harmed by depth of call tree by eliminating
all superfluous modifiers. For example, the pointer input
modifier does not need to be in the call chain for drawing
and draw modifiers don't need to be in the call chain for
pointer input.

Test: manual testing, small adjustment of broken tests.

Change-Id: I23b38aa0ed8b13fae603e336f14804c0b93e85f6