Add XMethodElement#name to reflect source name
This CL adds back the name property to XMethodElement. Unlike jvmName,
this property refers to the name in source and might be different from
the JVM name.

Note that, for functions that reference a value class, kotlin generates
a name with - in it, which makes it invalid in java source. As such,
these methods don't show up in KAPT stubs but they do show up in
compiled code. We should probably address that inconsistency in
XProcessing eventually (b/206517658).

A followup CL will update Room to use name where it makes sense to fix
the internals bug.

Bug: 206517658
Bug: 205289020
Test: XExecutableElementTest

Change-Id: I79864ffdde24d7edd7f0c49f668efc920ace3857