Introduce Renderer.SharedAssets
Bug: 202381485
Test: Manual testing and new unit tests
Relnote: Its common when editing a watch face for there to be both an interactive and a headless instance. To help save memory we've introduced Renderer.SharedAssets which allows a watch face renderer to share immutable data (e.g. textures and shaders) between instances. GlesRenderer.setEglConfig and GlesRenderer.setEglDisplay are depreciated, it was never intended for these to be settable, and doing so would have lead to undefined behavior.
Change-Id: I0d9e730bb712afc24dbd84611c6917f248772a27