Ensure PendingIntents with different params are different
The data for the Intent is set if it has parameters. The data contains
parameters to uniquely identify the location of the view.

As doing so requires a change in the definition of the activity, any
activity with parameters is now started using a trampoline. Only
activity from Intent are not modified, as the user can set the data
themselves if they need to.

Note: after this, we still need to handle more than one clickable action
on a given view.

Bug: 207612514
Relnote: n/a
Test: ./gradlew :glance:glance-appwidget:test

Change-Id: Ib1f15510bf11579135b85cadc3ccf5f1873808e1