Introduce ComplicationDataTimeline
Under the hood ComplicationDataTimeline is stored within a
ComplicationData's bundle, if we where starting from a green
field project this is not how we would have designed this but its
necessary given the need for backwards compatiability.

We could expose the timeline on the androidx ComplicationData
classes but chose not to because the resulting API would be
confusing. Instead wear-watchface transparently handles the
timeline overrides.

Note we also chose not to add this to
onSynchronousComplicationRequest becuase tha's really meant for
fast updates which is the opposite use case.

Test: Manual testing
Bug: 196958807
Relnote: We've added ComplicationDataTimeline to wear-complication-data-source this can be used to provide a sequence of time-gated ComplicationData to be delivered to the watch face which can be cached and updated automatically. E.g. today's weather forecast at various times or multiple upcoming calendar events. ComplicationRequestListener has been extended with a new method onComplicationDataTimeline which you can use to return this data.  There's a new kotlin wrapper SuspendingTimelineComplicationDataSourceService for suspending data source services.
Change-Id: Idecdc0c92a52ff9591b7a84e500986d4817719be