AndroidX Webkit: move loadUrl call into test body
I noticed while working on my last CL that loadUrl() was called as part
of the test setup (during Activity#onCreate), which meant that
@RepeatRule didn't actually repeat this part of the code. This change
moves the loadUrl() call into the test body, which is probably more
appropriate since this is the key functionality that's actually being
verified by the test case.

I'm not sure if this will improve flakiness, but I think this change is
probably sensible regardless.

Bug: 211597907
Test: verify HTML resources are requested for each test run
Change-Id: I0b111c767a179d733ea93ed35846b43a40cf6369