Throw proper exceptions for MutableVector List implementation
Fixes: 210875617

MutableVector has asMutableList() to return a List implementation.
We rely on using MutableVectors properly and don't check for
bounds, but List implementations are exposed to public API and
must throw the proper exceptions. This CL adds checks for index
bounds and throws the IllegalArgumentException and
IndexOutOfBoundsException as appropriate for index access methods.

Test: New tests
Change-Id: I55716a12b976682d2f111c82c69c9cb25b2d4625