Make ViewLayer work with Android Transitions
Fixes: 200421742
Fixes: 211485877

Android Transitions walk the view hierarchy looking for
any changes. ViewLayers can be added and removed during
drawing and shouldn't be controlled by the Android Transition
system as they are recycled internally. This CL prevents
transitions from walking the ViewLayer children by having
its parent return 0 from getChildCount(). Because ViewLayers
are treated essentially as RenderNodes, the only time the
children should be walked is during drawing. Though this
seems hacky, it is valid within the limited bounds of this
use.

Test: manual, new test
Change-Id: Ie6d73a53e308f31f3cbefda1b10928236cab7da6