Integrate CreationExtras into Fragment
Providing access to default CreationExtras from Fragment so they can be
used by custom ViewModelProvider factories.

Also did the same for the FragmentViewLifecycleOwner.

The error message before you the call to enableSavedStateHandles() is
currently not very helpful to devs since it is does not indicate that
you need to wait until onCreate(), but we will address that separately.

RelNote: "`Fragment` and `FragmentViewLifecycleOwner` now provide `CreationExtras` via the `getDefaultViewModelCreationExtras()` method."
Test: ViewModelTest, FragmentViewLifecycleOwnerTest
Bug: 207012585

Change-Id: I3060ba63bae3019489fc54eb116c3b4c1e50d8a1