Fix TextFieldKeyInput onValueChange
TextFieldKeyInput uses onValueChange that was provided by developer
which is not the correct callback to use.

state.onValueChange points to the developer provided callback. It does
not invalidate the composable. Therefore for cases where the text value
in not reset to a value (such as limiting to a character count) the
composable would not invalidate; the buffer would not be reset, and
still have the previously actually rejected input.

This CL changes TextFieldKeyInput to use the correct onValueChange
callback so that it invalidates CoreTextField and resets the editing
buffer making it aware of the rejection of the input.

Test: manual test on emulator / hardware keyboard
Test: HardwareKeyboardTest#textField_onValueChangeRecomposeTest

Bug: 200577798
Change-Id: I277ecbfb0be5d9a7f1f6536cacc3e467e8448149