ScalingLazyColumn.kt - add autoCentering property
Add a new property autoCentering: Boolean defaulted to true to ScalingLazyColumn.kt that will add space before the first and after the last list items in order to ensure that they can be scrolled to the center of the viewport. This avoids the need for a developer to try and work out how much contentPadding might be required on different sizes of wearables/lists so that all items can be scrolled to the center to be properly viewed.

RelNote: "We have added a new property autoCentering to the ScalingLazyColumn. When true (the default) this will ensure that all items including the first and last can be scrolled so that they are visible in the center of the lists viewport. Note that if using auto-centering you will probably want to set vertical content padding to 0.dp. If both autoCentering and vertical content padding are provided then they will both result in addition space being available before the first and after the last list items allowing them to be scrolled even further."

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214922490
Change-Id: I2a2827ad79ea6cf1cb5d9a623b600417def9cc2f