Add AbstractDecoysLowering and make all decoys lowerings extend it
This CL aligns the decoy lowerings with kotlin 1.6.0 changes to properly calculate signatures.
Decoy transformations rely on a proper signature calculation.
AbstractDecoysLowering implements the common logic needed for all decoy lowerings involving signature calculations.

This change is only for k/js targets (decoys are not used for jvm and native).

Test: tested in JB fork
Change-Id: Ifb68d5288d67444ee12fc3def2bc3fc0e7560ae0