Fix testSimpleVectorDrawables
On very recent Android framework versions, the "clip_filltype_nonzero"
case started failing, and I observed locally the difference is that the
edge antialiasing improved on the circle.

The simplest way to fix it is to delete the case and move the
fillType="nonZero" to the more robust test "clip_path_1". nonZero is
the default fillType anyway, so the only additional coverage needed
is that the string is parsed correctly.

Test: VectorDrawableTest.testSimpleVectorDrawables
Fixes: 213895983
Change-Id: I4ebbba8f71de4a6f21a3a026bad5050ab93cf062