Make sure `tapAction` is serialized.
We had a bug where the `tapAction` field was not being serialized for
the photo image complication.

This was not caught by our tests because we actually tested the one of
the data types in the test that was supposed to cover this class.

This commit updates the test to test the right class, fixes the bug,
and adds an additional class of tests for to make sure the data is
also parsed back, which was another bug that I detected by inspection.

Bug: 216050637
Relnote: "Fixes bug with PhotoImageComplicationData tapAction not
being correctly handled"
Test: ./gradlew :wear:watchface:watchface-complications-data:test

Change-Id: I1cc30a56e343c6f0b41406b3884e8b945f4085e8