Fix Java support for ListenableGlesRenderer
While this is technically an API breaking change for Java (kotlin
handles exceptions differently and is unaffected), it was not
possible to compile a class extending ListenableGlesRenderer
before this patch (see bug for details), making this a special
case.

Bug: 216267170
Test: A java class extending ListenableGlesRenderer actually compiles now.
Relnote: ListenableGlesRenderer's constructor is now correctly marked as @Throws(GlesException::class), and it is now possible to extend this class in java.
Change-Id: Iac6d0037a736c34e733622688cc3f68e9fa633c5