Update the ScalingLazyColumn scaling/alpha calculations to match latest understanding of UX Motion Spec
This patch corrects a misinterpretation of the UX Motion spec in terms of how to calculate the scaling transition point for each list item.
The previous calculation incorrectly assumed that the scaling line for an edge was between the edge and the center of the view port, this is incorrect and in fact the scaling line can actually be the other side of the view port center line.

Picker has had its default scaling parameters updated to take account for the different scaling characteristics.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 202164558
Change-Id: Ia8a44f2ce21c748f7251f999e1270f7c360d02bd