ScalingLazyColumn: change size semantics to match LazyColumn
Change ScalingLazyColumn default size to match LazyColumn by removing the Modifier.fillMaxSize() call on the contained LazyColumn .

RelNote: "We have modified ScalingLazyColumn so that it no longer greedily fills all of the space in its parent. Instead it will take its size from the size of its contents. This makes it consistent with the behaviour of LazyColumn. If you want to reinstate the old behaviour then pass Modifier.fillMaxSize() to the ScalingLazyColumn"
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214316016
Change-Id: I3cbfa038e71a6137d052fa5f41224df1563d5bd7