Allow AndroidViewModelFactory to create basic view models
When the stateless SavedStateViewModelFactory was added in aosp/1859226,
it unintentionally made it so that when using CreationExtras an
application was required to create a view model. That is not a hard
requirement of SavedStateViewModelFactory and this was caused by it
using AndroidViewModelFactory by default and that should also not be a
constraint there.

When creating a view model via AndroidViewModelFactory when using
CreationExtras, we should not require an application, and should instead
fall back to the NewInstanceFactory behavior.

RelNote: "`AndroidViewModelFactory` will not longer require an
application to be set when creating a view model via `CreationExtras`."
Test: AndroidViewModelFactoryTest
Bug: 217271656

Change-Id: I6ebefdf40940a08576a92ee4fb2dcc2b561d781d