Fix NavigationDrawer anchor on large screens.
NavigationDrawer in material3 has fixed width of the sheet, but the anchor to which theis sheet swipes is still set to fullscreen. This makes it unpleasant to open a drawer on a large screen, e.g. a tablet, as you see a scrim and no sheet until you reach the last bits of the screen.

This CL fixed this to set the anchor to the hardcoded drawerContent size.

Also, as a side effect, this makes drawer more performant, as the BoxWithConstraints is removed.

Fixes: 217353282
Test: adjusted, + screenshot tests need update
Change-Id: I5b2c17a0125b70f214f4b86085c6ab0c92f1f640