ScalingLazyColumn - reduce extraPadding from 10%->5% to avoid composing list items that are not displayed.
Change to reduce the amount of space allowed of extra list items that are composed in the underlying LazyList from 10% to 5%. The extra list items are composed in case they are needed as a result of the scaling down in size of other list components. Discussing with UX folks, doing some maths for standard scaling and list sizes we have determined we can adjust the extra padding down towards 0% for standard list scaling parameters and expected use cases. For now reduce to 5% and get some developer feedback on real world use cases. This behaviour can be overridden in the API if additional items are needed for a given scaling strategy.

RelNote: "We have reduced the ScalingLazyColumn default `extraPadding` that is provided to ensure that there are plenty of list items to draw (even when we are scaling some of them down in size) from 10% to 5%. This will avoid composing extra list items that might not appear in the viewport. If non standard scalingParams are being used (more extreme scaling for instance) the developer can adjust extra padding using viewportVerticalOffsetResolver."
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214993637
Change-Id: I76be4dff5489bc7c07e3a1c0b2857ba4edf9f6bd