Disable LayoutNodes reusing for nested SubcomposeLayouts
Recently SubcomposeLayout was converted to be ReusableComposeNode. It was incorrect as the reusing state from the parent composition is not being propagated to the nested composition in the current implementation. This causes nested subcompositions to not being recomposed with a new content. This change converts SubcomposeLayout back to be ComposeNode.

Fixes: 217365788
Fixes: 216833543
Test: new test in LazyColumnTest, plus had to remove the test which was testing now removed reusing logic
Change-Id: If9c8ee24d74ce39b5a82912259e55536a1d32a44