Add Placeholders to NoDataComplicationData
Bug: 196958807
Test: Manual testing + new unit tests
Relnote: Up to date ComplicationData may not always be avaliable (e.g. expired cached ComplicationData) so we've extended NoDataComplication with an optional placeholder ComplicationData and added ComplicationText.PLACEHOLDER, MonochromaticImage.PLACEHOLDER, SmallImage.PLACEHOLDER, PhotoImage.PLACEHOLDER which are only allowed to be used inside the context of a NoDataComplicationData placeholder. If selected these placeholders are suggested to be rendered with grey boxes/arcs.
Change-Id: I6285d065676350564f74292337d982dd8e52da01