[GH] Use programmatic configuration for build-cache cleanup
## Proposed Changes

A previous attempt to do this via a CLI argument failed, since the argument did
not apply to the cache settings for the `playground-plugin` included build.
Since the `playground-plugin` settings are loaded first, the CLI settings were
being ignored.

This change configures the cleanup threshold directly in Gradle settings, and does
so both for the Playground included build and the main build. This results in
any build-cache entries that have not been used in the previous day to be removed.

## Testing

Test: Tested by monitoring GitHub Actions pipeline

This is an imported pull request from https://github.com/androidx/androidx/pull/324.

Resolves #324
Github-Pr-Head-Sha: b3c8c679efff3c0f18b890da5edc8530b5b12db8
GitOrigin-RevId: 6e5ae554d34668bd84cda834fc240a83e3d9b524
Change-Id: I406484a2bd3147dbd7715a041666c1e4d244d044