Allow empty string args in deep links
An empty string arg is not the same as a null string arg. So we should
parse them and store the empty string.

This also makes sure we consider mismatch types as non matches. Ensuring
that we properly return a null matching bundle.

RelNote: "Empty string are now considered as value arguements in deep
links."
Test: NavDeepLinkTest
Bug: 217399862

Change-Id: I70a0d1b47db028d829014f21763f95bb76fa8e3c